except files that are required when clearing output folder

+++ src/app.py
@@ -151,13 +151,13 @@ def run_text_extractor_and_models(studyInstanceUID, description, prompt, output_
     # clear output folder textembedding
     for filename in os.listdir(FILES_FOLDER+"/text_embed"):
         file_path = os.path.join(FILES_FOLDER+"/text_embed", filename)
-        if os.path.isfile(file_path):
+        if os.path.isfile(file_path) and "dont_delete" not in filename:
             os.remove(file_path)
 
     # clear output folder low-resolution
     for filename in os.listdir(FILES_FOLDER +"/img_64_standard"):
         file_path = os.path.join(FILES_FOLDER +"/img_64_standard", filename)
-        if os.path.isfile(file_path):
+        if os.path.isfile(file_path) and "dont_delete" not in filename:
             os.remove(file_path)
 
     try:
